adds comparison module tes and makefile
