+++ src/nif/c_src/sort_nif.c
@@ -4,7 +4,7 @@ extern int quicksorthybrid(int* array, int l, int r);
 
 static ERL_NIF_TERM sort_nif(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[])
 {
-    int* input;
+    int input[10000];
     int N;
 
     if (!enif_get_list_length(env, argv[0], &N)) {
@@ -12,15 +12,18 @@ static ERL_NIF_TERM sort_nif(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]
     }
 
     int i = 0;
+    int C;
     ERL_NIF_TERM head, tail;
     ERL_NIF_TERM currList = argv[0];
     while(enif_get_list_cell(env, currList, &head, &tail)) {
-        if(!enif_get_int(env, head, input + i)) {
+        if(!enif_get_int(env, head, &C)) {
             return enif_make_badarg(env);
         }
+        input[i] = C;
         i++;
         currList = tail;
     }
+
 
     quicksorthybrid(input, 0, N-1);
