+++ src/ports/erl_comm.c
